+++ pos_discount_total/__manifest__.py
@@ -1,7 +1,6 @@
-# -*- coding: utf-8 -*-
 {
     'name': 'Discount for total amount of pos order',
-    'version': '10.0.1.0.0',
+    'version': '11.0.1.0.0',
     'author': 'IT-Projects LLC, Ivan Yelizariev',
     'license': 'LGPL-3',
     'category': 'Point Of Sale',