+++ pos_pin/static/src/js/pos.js
@@ -38,7 +38,7 @@ odoo.define('pos_pin.pos', function (require) {
                     });
                 }
             }
-    
+
             this.show_popup('selection',{
                 'title': options.title || _t('Select User'),
                 'list': list,
@@ -52,10 +52,15 @@ odoo.define('pos_pin.pos', function (require) {
 
             return def.then(function(cashier){
                 if (options.security && cashier !== options.current_user && cashier.pos_security_pin) {
-                    return self.ask_password(cashier.pos_security_pin, options.arguments);
+                    return self.ask_password(cashier.pos_security_pin, options.arguments).then(function(){
+                        return cashier;
+                    });
                 }
                 return cashier;
             }).done(function(res){
+                if (!res) {
+                    return;
+                }
                 return self.check_then_set_and_render_cashier(options, res);
             });
         },
@@ -81,12 +86,17 @@ odoo.define('pos_pin.pos', function (require) {
                     } else {
                         self.show_popup('error', {
                             'title': _t('Incorrect Password'),
-                            confirm: _.bind(self.show_password_popup, self, password, lock),
-                            cancel: _.bind(self.show_password_popup, self, password, lock),
+                            confirm: _.bind(self.show_password_popup, self, password, lock, cancel_function),
+                            cancel: _.bind(self.show_password_popup, self, password, lock, cancel_function),
                         });
                     }
                 },
-                cancel: cancel_function || lock.reject,
+                cancel: function() {
+                    if (cancel_function) {
+                        cancel_function.call(self);
+                    }
+                    lock.reject();
+                },
             });
             return lock;
         },