+++ pos_invoice_pay/doc/changelog.rst
@@ -1,8 +1,22 @@
-`1.0.3`
+`1.2.1`
 -------
 
 **Fix:** Don't auto-complete the current order after paying the invoice
 **Fix:** Incorrect receipt printing in the PosBox
+
+`1.2.0`
+-------
+
+**NEW:** Smart button on the pos session form to show paid regular invoices
+
+`1.1.1`
+-------
+
+**Fix:** Issue related to difference account
+
+`1.1.0`
+-------
+**Fix:** Ask cashier when click on fetch orders/invoices in POS
 
 `1.0.2`
 -------